Moving functions to dedicated files
