+++ data/products.go
@@ -1,13 +1,8 @@
 package data
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
-	"regexp"
 	"time"
-
-	"github.com/go-playground/validator"
 )
 
 // Product defines the structure for an API product.
@@ -23,37 +18,8 @@ type Product struct {
 	DeletedOn   string  `json:"-"`
 }
 
-func (p *Product) Validate() error {
-	validate := validator.New()
-	validate.RegisterValidation("sku", validateSKU)
-
-	return validate.Struct(p)
-}
-
-func validateSKU(fl validator.FieldLevel) bool {
-	// sku is of format abc-absd-dfsdf
-	re := regexp.MustCompile(`[a-z]+-[a-z]+-[a-z]+`)
-	matches := re.FindAllString(fl.Field().String(), -1)
-
-	if len(matches) != 1 {
-		return false
-	}
-
-	return true
-}
-
 // Products is a collection of Product
 type Products []*Product
-
-func (product *Product) FromJson(reader io.Reader) error {
-	decoder := json.NewDecoder(reader)
-	return decoder.Decode(product)
-}
-
-func (products *Products) ToJSON(w io.Writer) error {
-	encoder := json.NewEncoder(w)
-	return encoder.Encode(products)
-}
 
 // Returns the list of products
 func GetProducts() Products {
@@ -92,7 +58,7 @@ func getNextId() int {
 }
 
 // productList is a hard coded list of products for this
-// example data source
+// example data source. Should be replaced by database connection
 var productList = []*Product{
 	{
 		ID:          1,
